==> 02_Twitter_Sentiment_Analysis/notebooks/00_exploration.ipynb <==
TRAIN_PATH = "../data/raw/train.pkl"
TEST_PATH = "../data/raw/test.pkl"

# load packages and download data
import pandas as pd
import numpy as np

import nltk
nltk.download("stopwords")
nltk.download("twitter_samples")

from nltk.corpus import twitter_samples
from nltk.corpus import stopwords

## Read data:

positive_tweets = twitter_samples.strings("positive_tweets.json")
negative_tweets = twitter_samples.strings("negative_tweets.json")

## Running some explorations:

# showing the length of data
print(len(positive_tweets))
print(len(negative_tweets))

# Show some samples from positive and negative tweets
[print(x,'\n') for x in positive_tweets[:5]]

# Show some samples from positiveand negative tweets
[print(x,'\n') for x in negative_tweets[:5]]

## Now preparing train and test sets with targets:
**Split train and test with 80% and 20% percents**

train_pos = positive_tweets[:4000]
test_pos = positive_tweets[4000:]

train_neg = negative_tweets[:4000]
test_neg = negative_tweets[4000:]

X_train = np.array(train_pos + train_neg)
X_test = np.array(test_pos + test_neg)

print("length of X_train: ", X_train.shape)
print("length of X_test: ", X_test.shape)

**Now creating the target data**

y_train = np.append(np.ones((len(train_pos),1)), np.zeros((len(train_neg),1)), axis=0)
y_test = np.append(np.ones((len(test_pos),1)), np.zeros((len(test_neg),1)), axis=0)

print("y_train shape: ", y_train.shape)
print("y_test shape: " , y_test.shape)

## store data into a dataframe:

train_df = pd.DataFrame({"tweets":X_train,
                         "sentiment": y_train.reshape(-1,)})

test_df = pd.DataFrame({"tweets":X_test,
                         "sentiment": y_test.reshape(-1,)})

train_df.sample(5)

___
## Export data for furthur processing:

train_df.to_pickle(TRAIN_PATH)
test_df.to_pickle(TEST_PATH)
