==> 02_Twitter_Sentiment_Analysis/notebooks/01_preprocessing.ipynb <==
TRAIN_PATH = "../data/raw/train.pkl"
TEST_PATH = "../data/raw/test.pkl"

FINAL_TRAIN = "../data/processed/train_final.pkl"
FINAL_TEST = "../data/processed/test_final.pkl"

# Load packages
import numpy as np
import pandas as pd

import pickle

import sys
sys.path.append("../scripts")
from Utils import build_freqs, process_tweet

pd.set_option("display.max_colwidth", None)

___
## Functions:

def extract_features(tweet, freqs=features):
    """
    function used to transfrom tweet into its numerical representation
    based on the stored features
    INPUT:
        tweet: the tweet in a string format
        freqs: a dictionary of all the word frequency
                e.g {(great, 1): 35}
    OUTPUT:
        returns a numpy array of each score for features 
        e.g [bias, positive_score, negative_score]
    """
    words = process_tweet(tweet)
    tweet_array = np.array([[1.0, 0.0, 0.0]])
    
    for word in words:
        tweet_array[0, 1] += freqs.get((word, 1.0), 0)
        tweet_array[0, 2] += freqs.get((word, 0.0), 0)
    
    assert(tweet_array.shape == (1, 3))
    return tweet_array#.reshape(3,)

___
## Read data

train_df = pd.read_pickle(TRAIN_PATH)
test_df = pd.read_pickle(TEST_PATH)

train_df.sample(5)

test_df.sample(5)

## Now building features:
**From build_freqs fuction we will have a dictionary of {(words,label): freq} as our features**

features = build_freqs(X_train, y_train)
print(type(features))
print("length of features: ", len(features.keys()))

# Showing some of features
list(features.items())[:5]

**Now construct your features using extract_features function**
the output of shape `[[1], [0], [0]] `  will be splitted and used to construct a dataframe of three features; [bias, pos, neg]

train_features = train_df['tweets'].apply(extract_features)
train_features[:5]

As we can see the features are formated as a numpay array and so we will extract them to build our features as follows:
* bias: the bias term in the equation
* pos: the positive score of tweet
* neg: the negative score of the tweet

bias = [u[0,0] for u in train_features.to_numpy()]
pos = [u[0,1] for u in train_features.to_numpy()]
neg = [u[0,2] for u in train_features.to_numpy()]

X_train = pd.DataFrame({"bias": bias,
              "pos": pos,
              "neg": neg})

X_train.sample(5)

# Adding the target to final train features
X_train['sentiment'] = train_df['sentiment']
X_train.sample(5)

**Do the same thing with test data**

test_features = test_df['tweets'].apply(extract_features)
test_features[:5]

bias = [u[0,0] for u in test_features.to_numpy()]
pos = [u[0,1] for u in test_features.to_numpy()]
neg = [u[0,2] for u in test_features.to_numpy()]

X_test = pd.DataFrame({"bias": bias,
              "pos": pos,
              "neg": neg})

X_test.sample(5)

# Adding the target to final train features
X_test['sentiment'] = test_df['sentiment']
X_test.sample(5)

___
## Export features

X_train.to_pickle(FINAL_TRAIN)
X_test.to_pickle(FINAL_TEST)

with open("../data/objects/features.pkl", "wb") as handle:
    pickle.dump(features, handle)
