==> 02_Twitter_Sentiment_Analysis/notebooks/03_predicting.ipynb <==
FEATURES_DICT = "../data/objects/features.pkl"
MODEL = "../models/classifier.pkl"

# Load packages
import numpy as np
import pandas as pd

import pickle

import sys
sys.path.append("../scripts")
from Utils import process_tweet

___
## Functions:

with open(FEATURES_DICT, 'rb') as handle:
    features = pickle.load(handle)

def extract_features(tweet, freqs=features):
    """
    function used to transfrom tweet into its numerical representation
    based on the stored features
    INPUT:
        tweet: the tweet in a string format
        freqs: a dictionary of all the word frequency
                e.g {(great, 1): 35}
    OUTPUT:
        returns a numpy array of each score for features 
        e.g [bias, positive_score, negative_score]
    """
    words = process_tweet(tweet)
    tweet_array = np.array([[1.0, 0.0, 0.0]])
    
    for word in words:
        tweet_array[0, 1] += freqs.get((word, 1.0), 0)
        tweet_array[0, 2] += freqs.get((word, 0.0), 0)
    
    assert(tweet_array.shape == (1, 3))
    return tweet_array#.reshape(3,)

## Load model:

with open(MODEL, 'rb') as handel:
    model = pickle.load(handel)

def classify_tweet(tweet, clf):
    features = extract_features(tweet)
    
    X_train = pd.DataFrame({"bias": [features[0,0]],
              "pos": [features[0,1]],
              "neg": [features[0,2]]})
    y = clf.predict(X_train)
    
    return "Positive Sentiment" if y else "Negative Sentiment"

## Test on your own tweet:

**1. Try on a negative sentiment**

neg_tweets = ["Car &amp; suspemsion cant cope with all these offroading adventures now theres harvest machinery left in inapropriate places to weave around :((",
              
             "@marlenejazmyne idk :-( maybe because I think boobs are more fun to play with",
             
             'is everyone gonna be talking abt rat boy today bc :((',
             
              "@JackJackJohnson @jackgilinsky NUMBER ONE IN AMERICA!! 😭🇺🇸\ni wish i could buy it but no money :( http://t.co/7xyNmimTpR",
             
             "@denissely Awhh i didn't see you guys or i would've rammed my car into you :(",
             
             '@iperfectyonce @justinbieber its my biggest dream can u follow me brooo :((((',
             
             "@VyenAngel gosh its cheaper in malaysia :( here its worth 130 +shipping",
             
             "@Miss_J_Hart @staffrm well, I'm only on P.41 &amp; 8 pages of notes thus far :-( and I've read twice before",
             
             ":( I can no longer admire the consistency of Jason Shackell's hair on the @NoNayNeverNet podcast http://t.co/T9rpuLyb5u",
             
             "@voxcinemas I want the minion bucket please :( it wasn't there yesterday! :("
             ]

[(i, classify_tweet(tweet, model)) for i, tweet in enumerate(neg_tweets)]

**2. Try on a positive sentiment**

pos_tweets = ["Just viewed a property. Agent said 'the kitchen needs attention'  I like an accurate description :-) http://t.co/kgXuYb6diH",
            
            'Deepthroat was a good movie :)) https://t.co/rtz4SsXA0O',
            
            "@PuppyShogun mistakes happen man, as long as we get to play the game, we'll be happy :)",
            
            '@monkeymademe the @raspjamberlin is still on tomorrow? A work colleague is interested in bringing his kid along :) can point him to EB page',
            
            '@mehra_sachi sure :)',
            
            '@IGNUK #1TBPS4 Wow! what a prize! so delightful for my eyes, pick me to win and I will burst and sing "PS4 is mine and I feel divine" :)',
             
             "@CathsStrawberry I thought your work was seasonal, you will have really busy days. Ok so deal for weekends and in your toilet breaks :p",
             
             "@iansomerhalder here in Spain is morning :) Good morning Ian!",
             
             "@Caz_Loz @BeckieSmith_7 ha! Love both pics! Clearly I have much to learn about popular English pastimes. :)",
             
             "@OMGitsBern_92 All sorted in the early hours. Thanks again for your help &amp; raising the initial issue. makes the community what it is :)"
             ]

[(i, classify_tweet(tweet, model)) for i, tweet in enumerate(pos_tweets)]
